# src/swagger/swagger_output.yaml
...
openapi: 3.0.0
info:
  version: 0.0.0
  title: Vinymatic API
  description: ""
host: localhost:5000
basePath: ""
...
schemes: []
securityDefinitions: {}
consumes: []
produces: []
paths:
...
  "/users":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/pageParam"
      - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Users"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    post:
      tags:
      - User
      description: ""
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/AddUser"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
...
  "/is/pseudo/available/{username}":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/userParam"
      responses:
        "200":
          description: OK
          content:
            text/plain:
              example: true
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    patch:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/requestBodies/AddUser"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "304":
          description: Not Modified
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    delete:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/username":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/playlists":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaylistOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/followings":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowingsOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/followers":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowersOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/collection":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/user/{id}/wishlist":
    get:
      tags:
      - User
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistOfUser"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/wishlist/vinyl/{id}/user/{userId}":
    get:
      tags:
      - User
      description: ""
      parameters:
      - name: id
        in: path
        require: true
        schema: 
          type: integer
      - name: userId
        in: path
        require: true
        schema: 
          type: integer
      responses:
        "200":
          description: Created

        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/collection/vinyl/{id}/user/{userId}":
    get:
      tags:
      - User
      description: ""
      parameters:
      - name: id
        in: path
        require: true
        schema: 
          type: integer
      - name: userId
        in: path
        require: true
        schema: 
          type: integer
      responses:
        "200":
          description: Created

        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
# Video ✔︎
...
  "/vinyl/{id}/formats":
    get:
      tags:
      - Vinyl
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
...
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    userParam:
      name: username
      in: path
      required: true
      schema:
        type: string
...
  requestBodies:
...
    AddUser:
      allOf:
      - type: object
        properties:
          username:
            type: string
          email:
            type: string
          phoneNumber:
            type: string
          firstName:
            type: string
          lastName:
            type: string
          showName:
            type: string
          birthDate:
            type: string
          note:
            type: integer
          localisation:
            type: string
          description:
            type: string
          password:
            type: string
        required:
        - username
        - password
      - type: object
        properties: 
          profilImage:
            type: string
            format: binary
          coverImage:
            type: string
            format: binary
...
  schemas:
...
    MinOfPlaylist:
      type: object
      properties:
        id: 
          type: integer
        title: 
          type: string
        duration: 
          type: integer
        image: 
          type: string
        playlistUrl: 
          type: string
        resourceUrl: 
          type: string
...
    MinOfUserForEntity:
      type: object
      properties:
        id: 
          type: integer
        username:
          type: string
    MinOfUser:
      type: object
      properties:
        id: 
          type: integer
        username:
          type: string
        email:
            type: string
        phoneNumber:
            type: string
        showName:
          type: string
        # firstName:
        #   type: string
        # lastName:
        #   type: string
        profilImage:
          type: string
        role:
          type: string  
...
    MinOfVinyl:
      type: object
      properties:
        id: 
          type: integer
        idRelease: 
          type: integer  
        title:
          type: string  
        thumbnail:
          type: string  
        vinylUrl:
          type: string
        resourceUrl:
          type: string
  
  # Array of
...
    ArrayOfPlaylist:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfPlaylist"
...
    ArrayOfUser:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfUser"
    ArrayOfVinyl:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfVinyl"
...
    Collection: 
      type: object
      properties: 
        Vinyls:
          allOf:
          - $ref: "#/components/schemas/ArrayOfVinyl"
          - type: array
            items:
              type: object
              properties: 
                VinylsInCollection:
                  type: object
                  properties:
                    coverCondition:
                      type: string
                    diskCondition:
                      type: string  
...
    User:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUser"
          - type: object
          properties:
            # email:
            #   type: string
            # phoneNumber:
            #   type: string
            firstName:
              type: string
            lastName:
              type: string
            birthDate:
              type: string
            note:
              type: string
            localisation:
              type: string
            description:
              type: string
            coverImage:
              type: string
            createdAt:
              type: string
            updatedAt:
              type: string
            followings:
              type: integer
            followers:
              type: integer
    UsernameOfUser:
      type: object
      properties:
        username:
          type: string
    PlaylistOfUser:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUserForEntity"
          - type: object
            properties:
              PlayList: 
                $ref: "#/components/schemas/ArrayOfPlaylist"
...
    FollowingsOfUser:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUserForEntity"
          - type: object
            properties:
              followings:
                type: integer
              Followings: 
                $ref: "#/components/schemas/ArrayOfUser"
    FollowersOfUser:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUserForEntity"
          - type: object
            properties:
              followers:
                type: integer
              Followers: 
                $ref: "#/components/schemas/ArrayOfUser"
    CollectionOfUser:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUserForEntity"
          - type: object
            properties:
              id:
                type: integer
              Collection:
                $ref: "#/components/schemas/Collection"
    WishlistOfUser:
      type: object
      properties: 
        user:
          allOf:
          - $ref: "#/components/schemas/MinOfUserForEntity"
          - type: object
            properties:
              id:
                type: integer
              Wishlist:
                $ref: "#/components/schemas/Wishlist"
...
    Wishlist: 
      type: object
      properties: 
        Vinyls:
          allOf:
          - $ref: "#/components/schemas/ArrayOfVinyl"
          - type: array
            items:
              type: object
              properties: 
                VinylsInWishlist:
                  type: object
                  properties:
                    coverCondition:
                      type: string
                    diskCondition:
                      type: string  
    # Own
...
    Users:
      type: object
      properties:
        users:
          allOf:
          - $ref: "#/components/schemas/ArrayOfUser"
...